langchain_examples/requirements.txt
# LangChain 基础教程 - 依赖

# LangChain 核心库
langchain>=0.3.0
langchain-community>=0.3.0

# LLM 集成
langchain-openai>=0.2.0
langchain-ollama>=0.2.0

# 数据处理
pydantic>=2.0.0

# 环境变量管理
python-dotenv>=1.0.0

# 文本嵌入（用于 RAG）
langchain-text-splitters>=0.3.0
